Webserver/app/requirements.txt
alembic==1.13.1
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.2
gunicorn==21.2.0
idna==3.6
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.1
